aggregator: move construction to generic controlplane
